python/adhoc/requirements.txt
gevent==24.2.1
psycopg2-binary
jinja2
